Fully functional rotation in different directions triggered by arrow buttons

+++ 3D rendering/Vertex.h
@@ -4,7 +4,7 @@
 
 
 void drawShape(SDL_Renderer* renderer, Cube shape);
-void rotateShapeX(int theta, Cube &shape);
-void rotateShapeY(int theta, Cube& shape);
-void rotateShapeZ(int theta, Cube& shape);
+void rotateShapeX(double theta, Cube &shape);
+void rotateShapeY(double theta, Cube& shape);
+void rotateShapeZ(double theta, Cube& shape);
 void applyMatrix(double transMatrix[3][3], Cube& shape);
